Move easy receipts handling to transact view
