+++ packages/lightwallet/mobile/src/app/wallets/wallets.ts
@@ -6,7 +6,7 @@ import { EasyReceiveService } from '@merit/common/services/easy-receive.service'
 import { LoggerService } from '@merit/common/services/logger.service';
 import { ProfileService } from '@merit/common/services/profile.service';
 import { MeritToastController, ToastConfig } from '@merit/common/services/toast.controller.service';
-import { Events, AlertController, IonicPage, NavController, Platform } from 'ionic-angular';
+import { Events, IonicPage, NavController, Platform } from 'ionic-angular';
 
 @IonicPage()
 @Component({
@@ -26,10 +26,8 @@ export class WalletsView {
 
   constructor(private navCtrl: NavController,
               private logger: LoggerService,
-              private easyReceiveService: EasyReceiveService,
               private toastCtrl: MeritToastController,
               private profileService: ProfileService,
-              private alertController: AlertController,
               private platform: Platform,
               private events: Events
   ) {
@@ -50,8 +48,6 @@ export class WalletsView {
       this.refreshing = true;
       this.updateAllInfo().then(() => this.refreshing = false);
     });
-
-    this.processPendingEasyReceipts();
   }
 
   private get isActivePage(): boolean {
@@ -129,186 +125,6 @@ export class WalletsView {
     this.totalAmount = totalAmount;
   }
 
-
-  /**
-   * gets easyReceipt data from the blockchain and routes the ui accordingly
-   *
-   * @private
-   * @param {EasyReceipt} receipt
-   * @param {boolean} isRetry
-   * @param {string} password
-   * @returns {Promise<void>}
-   * @memberof WalletsView
-   */
-  private async processEasyReceipt(receipt: EasyReceipt, isRetry: boolean, password: string = '') {
-    const data = await this.easyReceiveService.validateEasyReceiptOnBlockchain(receipt, password);
-    let txs = data.txs;
-
-    if (!txs) return await this.easyReceiveService.deletePendingReceipt(receipt);
-
-    if (!Array.isArray(txs)) {
-      txs = [txs];
-    }
-
-    if (!txs.length) return this.showPasswordEasyReceivePrompt(receipt, isRetry);
-
-    if (txs.some(tx => tx.spent)) {
-      this.logger.debug('Got a spent easyReceipt. Removing from pending receipts.');
-      await this.easyReceiveService.deletePendingReceipt(receipt);
-      await this.showSpentEasyReceiptAlert();
-      return await this.processPendingEasyReceipts();
-    }
-
-    if (txs.some(tx => (tx.confirmations === undefined))) {
-      this.logger.warn('Got easyReceipt with unknown depth. It might be expired!');
-      return this.showConfirmEasyReceivePrompt(receipt, data);
-    }
-
-    if (txs.some(tx => receipt.blockTimeout < tx.confirmations)) {
-      this.logger.debug('Got an expired easyReceipt. Removing from pending receipts.');
-      await this.easyReceiveService.deletePendingReceipt(receipt);
-      await this.showExpiredEasyReceiptAlert();
-      return await this.processPendingEasyReceipts();
-    }
-
-    return this.showConfirmEasyReceivePrompt(receipt, data);
-  }
-
-  /**
-   * checks if pending easyreceive exists and if so, open it
-   */
-  private async processPendingEasyReceipts(): Promise<any> {
-    const receipts = await this.easyReceiveService.getPendingReceipts();
-    if (!receipts.length) return; // No receipts to process
-    return this.processEasyReceipt(receipts[0], false);
-  }
-
-  /**
-   * Shows easy receive popup with password input. Input is highlighted red when password is incorrect
-   * @param receipt
-   * @param highlightInvalidInput
-   */
-  private showPasswordEasyReceivePrompt(receipt: EasyReceipt, highlightInvalidInput = false) {
-    this.logger.info('show alert', highlightInvalidInput);
-    this.alertController.create({
-      title: `You've got merit from ${receipt.senderName}!`,
-      cssClass: highlightInvalidInput ? 'invalid-input-prompt' : '',
-      inputs: [{ name: 'password', placeholder: 'Enter password', type: 'password' }],
-      buttons: [
-        {
-          text: 'Ignore', role: 'cancel', handler: () => {
-            this.logger.info('You have declined easy receive');
-            this.easyReceiveService.deletePendingReceipt(receipt).then(() =>
-              this.processPendingEasyReceipts()
-            );
-          }
-        },
-        {
-          text: 'Validate', handler: (data) => {
-            if (!data || !data.password) {
-              this.showPasswordEasyReceivePrompt(receipt, true); //the only way we can validate password input by the
-                                                                 // moment
-            } else {
-              this.processEasyReceipt(receipt, true, data.password);
-            }
-          }
-        }
-      ]
-    }).present();
-  }
-
-  /**
-   * Shows after receipt is validated with correct password (or with no pass if no one set)
-   * @param receipt
-   * @param data
-   */
-  private async showConfirmEasyReceivePrompt(receipt: EasyReceipt, data) {
-
-    let amount = await this.easyReceiveService.getReceiverAmount(data.txs);
-
-    this.alertController.create({
-      title: `You've got ${amount} Merit!`,
-      buttons: [
-        {
-          text: 'Reject', role: 'cancel', handler: () => {
-            this.rejectEasyReceipt(receipt, data).then(() =>
-              this.processPendingEasyReceipts()
-            );
-          }
-        },
-        {
-          text: 'Accept', handler: () => {
-            this.acceptEasyReceipt(receipt, data).then(() =>
-              this.processPendingEasyReceipts()
-            );
-          }
-        }
-      ]
-    }).present();
-  }
-
-  private showSpentEasyReceiptAlert() {
-    this.alertController.create({
-      title: 'Uh oh',
-      message: 'It seems that the Merit from this link has already been redeemed!',
-      buttons: [
-        'Ok'
-      ]
-    }).present();
-  }
-
-  private showExpiredEasyReceiptAlert() {
-    this.alertController.create({
-      title: 'Uh oh',
-      subTitle: 'It seems that this transaction has expired. ',
-      message: 'The Merit from this link has not been lost! ' +
-      'You can ask the sender to make a new transaction.',
-      buttons: [
-        'Ok'
-      ]
-    }).present();
-  }
-
-  private async acceptEasyReceipt(receipt: EasyReceipt, data: any): Promise<any> {
-    try {
-      const wallets = await this.profileService.getWallets();
-      // TODO: Allow a user to choose which wallet to receive into.
-      let wallet = wallets[0];
-      if (!wallet) throw 'no wallet';
-
-      const address = wallet.getRootAddress();
-      const acceptanceTx = await this.easyReceiveService.acceptEasyReceipt(receipt, wallet, data, address.toString());
-
-      this.logger.info('accepted easy send', acceptanceTx);
-      return this.updateAllInfo();
-    } catch (err) {
-      console.log(err);
-      this.toastCtrl.create({
-        message: 'There was an error retrieving your incoming payment.',
-        cssClass: ToastConfig.CLASS_ERROR
-      }).present();
-    }
-  }
-
-  private rejectEasyReceipt(receipt: EasyReceipt, data): Promise<any> {
-
-    return this.profileService.getWallets().then((wallets) => {
-
-      //todo implement wallet selection UI
-      let wallet = wallets[0];
-      if (!wallet) return Promise.reject(new Error('Could not retrieve wallet.'));
-
-      return this.easyReceiveService.rejectEasyReceipt(wallet, receipt, data).then(() => {
-        this.logger.info('Easy send returned');
-      }).catch((err) => {
-        console.log(err);
-        this.toastCtrl.create({
-          message: err.text || 'There was an error rejecting the Merit',
-          cssClass: ToastConfig.CLASS_ERROR
-        }).present();
-      });
-    });
-  }
 
 }
